requirements.txt
pytest
Flask==2.3.2
password-validator==1.0
blinker==1.6.2
click==8.1.6
colorama==0.4.6
exceptiongroup==1.1.2
Flask-WTF==1.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
tomli==2.0.1
Werkzeug==2.3.6
WTForms==3.0.1
SQLAlchemy==1.4.41
python-dotenv==0.19.0

